Refactor: move partner-level classes into their own file
Part of the ongoing breakup of binding.py
